+++ test/FunctionMetadataGeneratorTests/ExtensionsCsProjGeneratorTests.cs
@@ -200,7 +200,7 @@ namespace Microsoft.Azure.Functions.SdkTests
             return version switch
             {
                 FuncVersion.V3 => new ExtensionsCsprojGenerator(extensions, outputPath, "v3", Constants.NetCoreApp, Constants.NetCoreVersion31),
-                FuncVersion.V4 => new ExtensionsCsprojGenerator(extensions, outputPath, "v4", Constants.NetCoreApp, Constants.NetCoreVersion6),
+                FuncVersion.V4 => new ExtensionsCsprojGenerator(extensions, outputPath, "v4", Constants.NetCoreApp, Constants.NetCoreVersion8),
                 _ => throw new ArgumentOutOfRangeException(nameof(version)),
             };
         }
@@ -244,7 +244,7 @@ namespace Microsoft.Azure.Functions.SdkTests
             return @"
 <Project Sdk=""Microsoft.NET.Sdk"">
     <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
+        <TargetFramework>net8.0</TargetFramework>
         <AssemblyName>Microsoft.Azure.Functions.Worker.Extensions</AssemblyName>
         <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
     </PropertyGroup>
@@ -259,7 +259,7 @@ namespace Microsoft.Azure.Functions.SdkTests
 
     <Target Name=""_VerifyTargetFramework"" BeforeTargets=""Build"">
         <!-- It is possible to override our TFM via global properties. This can lead to successful builds, but runtime errors due to incompatible dependencies being brought in. -->
-        <Error Condition=""'$(TargetFramework)' != 'net6.0'"" Text=""The target framework '$(TargetFramework)' must be 'net6.0'. Verify if target framework has been overridden by a global property."" />
+        <Error Condition=""'$(TargetFramework)' != 'net8.0'"" Text=""The target framework '$(TargetFramework)' must be 'net8.0'. Verify if target framework has been overridden by a global property."" />
     </Target>
 </Project>
 ";